reverted factory from curriculum tracking to core + removed redundant import
