+++ backend/core/tests/factories.py
@@ -11,8 +11,17 @@ def _email_generator():
         yield f"foo.{i}@example.com"
         i += 1
 
+
+def _team_name_generator():
+    i = 1
+    while True:
+        yield f"GROUP NAME {i}"
+        i += 1
+
+
 _email_iterator = _email_generator()
 
+_team_name_iterator = _team_name_generator()
 
 
 class UserFactory(DjangoModelFactory):
@@ -25,6 +34,14 @@ class UserFactory(DjangoModelFactory):
 
     is_superuser = False
     is_staff = False
+
+
+class TeamFactory(DjangoModelFactory):
+    class Meta:
+        model = "core.Team"
+
+    name = factory.lazy_attribute(lambda *args, **kwargs: next(_team_name_iterator))
+    active = True
 
 
 class CurriculumFactory(DjangoModelFactory):
