+++ backend/curriculum_tracking/tests/test_command_export_project_urls.py
@@ -1,6 +1,5 @@
 from django.test import TestCase
-from core.tests.factories import UserFactory
-from curriculum_tracking.tests.factories import TeamFactory
+from core.tests.factories import TeamFactory, UserFactory
 from curriculum_tracking.tests.factories import ProjectContentItemFactory, RecruitProjectFactory
 from curriculum_tracking.management.commands.export_project_urls import get_user_and_projectlink
